[OpenACC][NFCI] Split recipe generation into its own class
The recipe generation was dependent on the clause kind, which meant we
had all of the recipe generation duplicated in each of clauses.  This
patch copy/pastes all of them into their own type to do recipe
generation, which should reduce clang's size.

Additionally, we've moved it off into its own file, which should make
readability/organization improvements.
